Drop ver.py source file and use python version macros
The %{python2_version} and %{python3_version} macros are pre-defined, so
we don't need to use ver.py to find the versions.

Use shell variable for Python 3 ABI flags instead of global macro. This
avoids errors when creating SRPMs or running rpmlint, because by
delaying the command until the %build stage we can rely on python3-devel
being installed.