Disable openmpi and mpich for Flatpak-bundled builds
The Fedora OpenMPI and MPICH Flatpaks pose considerable challenges for
rebuilding to use in Flaptaks with prefix=/app - in particular because
of their use of environment-modules. An analysis of all graphical
applications in Fedora that we might want to create Flatpak containers
of shows no apps that use the OpenMPI and MPICH subpackages of boost,
though many other boost packages are used. So simply disabling openmpi
and mpich for Flatpak rebuilds is the simplest approach.